ajout florent

+++ src/gilded/roses/objects/GrowingItem.java
@@ -2,7 +2,7 @@ package gilded.roses.objects;
 
 public class GrowingItem extends Item {
 
-	public GrowingItem(String name, short sellIn, short quality) {
+	public GrowingItem(String name, int sellIn, int quality) {
 		setName(name);
 		setSellIn((sellIn >= 0 ? sellIn : 0));
 		setQuality((quality >= Item.qualiteMini ? (quality <= Item.qualiteMax ? quality : Item.qualiteMax) : Item.qualiteMini));
@@ -11,9 +11,9 @@ public class GrowingItem extends Item {
 	@Override
 	public void newDay() {
 		if (getSellIn() > 0)
-			setSellIn((short) (getSellIn()-Item.valeurDescendante));
+			setSellIn((int) (getSellIn()-Item.valeurDescendante));
 
-		setQuality((short) (getQuality() + Item.valeurMontante <= Item.qualiteMax ?
+		setQuality((int) (getQuality() + Item.valeurMontante <= Item.qualiteMax ?
 				getQuality()+Item.valeurMontante : Item.qualiteMax));
 	}
